requirements.txt
gradio
pydantic
python-dotenv
google-genai
splitwise